requirements.txt
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
black==19.10b0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
dj-database-url==0.5.0
Django==3.0
django-extra-fields==2.0.1
django-filter==2.2.0
django-heroku==0.3.1
djangorestframework==3.10.3
drf-yasg==1.17.0
gunicorn==20.0.4
h11==0.8.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
inflection==0.3.1
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==19.2
pathspec==0.6.0
Pillow==6.2.1
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyHamcrest==1.9.0
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
pyOpenSSL==19.1.0
pyparsing==2.4.5
pytz==2019.3
regex==2019.11.1
requests==2.22.0
ruamel.yaml==0.16.5
service-identity==18.1.0
six==1.13.0
sqlparse==0.3.0
toml==0.10.0
Twisted==19.10.0
txaio==18.8.1
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.7
uvicorn==0.10.8
websockets==8.1
whitenoise==4.1.4
wrapt==1.11.2
zope.interface==4.7.1
